Add files via upload
This project is a simple countdown timer implemented using HTML, CSS, and JavaScript. It allows users to set a specific date and time in the future and count down the time remaining until that event occurs.

Features
Customizable Timer: Users can specify the target date and time for the countdown, allowing for flexibility in tracking various events and deadlines.
Real-time Updates: The timer updates in real-time, displaying the remaining days, hours, minutes, and seconds until the specified event occurs.
User-Friendly Interface: The interface is designed to be intuitive and user-friendly, making it easy for users to set up and interact with the countdown timer.
Responsive Design: The countdown timer is responsive and adapts to different screen sizes and devices, ensuring a consistent experience across desktop, tablet, and mobile platforms.
Custom Styling Options: The codebase includes customizable styling options, allowing developers to adjust the appearance of the countdown timer to suit their project's design and branding.
Usage
To use the Countdown Timer, clone this repository to your local machine and open the index.html file in a web browser. Set the target date and time for the countdown, and the timer will automatically start counting down.

Credits
This project was developed by Muhammed A. Suleiman